# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing user carts and orders.",
        "title": "Cart API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/cart": {
            "get": {
                "description": "Возвращает корзину пользователя со списком товаров",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Получить корзину пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "страница пагинации",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "лимит пагинации",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Cart"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет товар с указанным id и количеством в корзину пользователя и возвращает обновленную корзину. В случае, если в корзине уже есть товар с данным id - увеличит количество товара в корзине на указанное число.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Добавить товар в корзину",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "description": "Данные о товаре",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cart.addDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Cart"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Изменяет количество указанного товара в корзине пользователя. Не добавляет товар, если его нет в корзине",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Изменить количество товара в корзине",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "description": "ID товара и новое количество",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cart.quantityDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cart/{product_id}": {
            "delete": {
                "description": "Удаляет товар с указанным ID из корзины пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Удалить товар из корзины",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID товара",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Cart"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/order": {
            "get": {
                "description": "Возвращает список всех заказов пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Получить заказы пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "completed",
                        "description": "сортировка по статуса",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "страница пагинации",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "лимит пагинации",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Order"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новый заказ основываясь на содержимом корзины",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Создать заказ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/order/status": {
            "patch": {
                "description": "Изменяет статус указанного заказа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Изменить статус заказа",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "description": "ID заказа и новый статус",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order.statusDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/order/{orderID}": {
            "get": {
                "description": "Возвращает заказ с указанным ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Получить заказ по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID заказа",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cart.addDTO": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "cart.quantityDTO": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "quantity": {
                    "type": "integer",
                    "example": 4
                }
            }
        },
        "entity.Cart": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CartItem"
                    }
                },
                "totalQuantity": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "entity.CartItem": {
            "type": "object",
            "properties": {
                "cart_id": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-05-05 10:30:00"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "quantity": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "entity.Order": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "cost": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.OrderItem"
                    }
                },
                "itemsQuantity": {
                    "type": "integer"
                },
                "rawCost": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "entity.OrderItem": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "orderID": {
                    "type": "integer"
                },
                "productID": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "rawCost": {
                    "type": "number"
                }
            }
        },
        "entity.OrderStatus": {
            "type": "string",
            "enum": [
                "created",
                "in_progress",
                "cancelled",
                "completed"
            ],
            "x-enum-varnames": [
                "CREATED",
                "IN_PROGRESS",
                "CANCELED",
                "COMPLETED"
            ]
        },
        "order.statusDTO": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.OrderStatus"
                        }
                    ],
                    "example": "completed"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}